requirements.txt
Django==1.7.7
Markdown==2.6.1
PyYAML==3.11
django-filter==0.9.2
django-rest-swagger==0.2.9
djangorestframework==3.1.1
factory-boy==2.5.1
